// src/redux/tag-types.ts
export enum tagTypes {
  user = "user",
  updateProfile = "updateProfile",
  updatePassword = "updatePassword",
  admin = "admin",
  blog = "blog",
  category = "category",
  service = "service",
  faq = "faq",
  servey = "servey",
  cart = "cart",
  comment = "comment",
  review = "review",
  booking = "booking",
}

export const tagTypesList = [
  tagTypes.user,
  tagTypes.updateProfile,
  tagTypes.updatePassword,
  tagTypes.admin,
  tagTypes.blog,
  tagTypes.category,
  tagTypes.service,
  tagTypes.faq,
  tagTypes.servey,
  tagTypes.cart,
  tagTypes.comment,
  tagTypes.review,
  tagTypes.booking,
];